+++ changelog.md
@@ -5,7 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-- None
+### Added
+- `AdvancedEffect` and `AdvancedEffectTrigger` have been added, and have a `Trigger` function that takes a parameter of the type you specify, which allows the effect to be triggered with additional context information.
+    - This is equivalent to the old method of passing parameters which entailed creating a subclass of `EffectArgs`.
+    - These classes accept a parameter of an arbitrary type, rather than forcing you to subclass `EffectArgs` to pass parameters.
+
+### Changed
+- All classes related to the effects system have been moved to the `GoRogue.Effects` namespace.
+- `Effect` and `EffectTrigger` no longer accept type parameters
+    - Versions called `AdvancedEffect` and `AdvancedEffectTrigger` have been added which do accept type parameters
+- Cancellation of a trigger from an effect is now handled via an `out bool` parameter given to `Trigger` and `OnTrigger`
+    - Set this boolean value to true to cancel the trigger
+
+### Removed
+- `EffectArgs` has been removed.
+    - Parameters to `AdvancedEffect` and `AdvancedEffectTrigger` can now be of an arbitrary type.
 
 ## [3.0.0-beta07] - 2023-07-11
 ### Added